+++ app/helpers/events_helper.rb
@@ -1,5 +1,10 @@
 module EventsHelper
   def attending?(event)
-  
+
   end
+
+  def creator(event)
+    creator = User.find(event.creator_id)
+  end
+  
 end
