+++ app/helpers/users_helper.rb
@@ -1,0 +1,15 @@
+module UsersHelper
+
+  def get_past_events(user)
+    past_event = []
+    Event.previous.each { |event| past_event << event if event.guests.ids.include? user }
+    past_event
+  end
+
+  def get_future_events(user)
+    future_event = []
+    Event.future.each { |event| future_event << event if event.guests.ids.include? user }
+    future_event
+  end
+
+end
