+++ app/views/events/index.html.erb
@@ -1,9 +1,7 @@
-Future events:
+<h5>Next events:</h5>
 
-Event.future.each do |event|
-list of future events
-end
+<%= render 'future' %>
 
-Past events:
-Event.previous.each do |event|
-end
+<h5>Past events</h5>
+
+<%= render 'past' %>
