+++ app/views/events/show.html.erb
@@ -1,6 +1,6 @@
 <h2>Event</h2>
 <h3>Creator:</h3>
-<%= User.find(@event.creator_id).name %>
+<%= link_to "#{creator(@event).name}", user_show_path(creator(@event)) %>
 <h3> Location: </h3>
 <%= @event.location %>
 <h3>Description:</h3>
@@ -8,6 +8,8 @@
 <h3>Date:</h3>
 <%= @event.date %>
 <h3>Attendee:</h3>
+<ul>
 <% @guests.each do |guest| %>
-  <p> <%= " * #{guest.name}" %> </p>
+  <li> <%= link_to "#{guest.name}", user_show_path(guest)  %> </li>
 <% end %>
+</ul>
