+++ app/views/layouts/application.html.erb
@@ -13,14 +13,15 @@
   <nav class="navbar navbar-inverse navbar-dark bg-dark mb-4">
   <div class="container-fluid">
     <div class="navbar-header">
-      <a class="navbar-brand" href="#">Private Events</a>
+      <%= link_to 'Private Events', root_path, class: "navbar-brand btn btn-lg btn-outline-light" %>
     </div>
     <ul class="nav navbar-nav">
-      <li class="">
+      <li class="d-inline-flex align-items-center ">
 
       <% if logged_in? %>
-      <h5 class="text-light d-inline-block mr-5" > <%= "Hello, #{current_user.name}"%></h5>
-      <%= link_to 'Log Out', logout_path, method: :delete, class: "btn btn-lg btn-outline-light " %>
+      <h5> <%= link_to "Hello, #{current_user.name}", user_show_path(current_user), class: "text-light mr-3" %></h5>
+      <%= link_to 'New event', create_event_path, class: "btn btn-lg btn-outline-light mr-3 " %>
+      <%= link_to 'Log Out', logout_path, method: :delete, class: "btn btn-lg btn-outline-light" %>
       <% else %>
        <%= link_to 'Log In', login_path, class: "btn btn-lg btn-outline-light " %>
       <% end %>
