+++ app/views/users/_future.html.erb
@@ -1,0 +1,11 @@
+<div class="p-4">
+<% @future_events.each do |inv| %>
+<div class="card" style="width: 18rem;">
+  <div class="card-body">
+  <h5 class="card-title"><%= "* You were invited to an event in #{inv.location } " %> </h5>
+  <p class="card-text"> <%= "at #{inv.date.to_formatted_s(:rfc822)}.  Will you go?" %></p>
+  <%= link_to 'View event', event_show_path(inv)  %>
+  </div>
+</div>
+<% end %>
+</div>
