+++ app/views/users/_past.html.erb
@@ -1,0 +1,15 @@
+<div class="p-4">
+<% if @past_events.empty? %>
+  <%= "There's not previous events" %>
+<% else %>
+<% @past_events.each do |inv| %>
+<div class="card" style="width: 18rem;">
+  <div class="card-body">
+  <h5 class="card-title"><%= "* You were invited to an event in #{inv.location } " %> </h5>
+  <p class="card-text"> <%= "at #{inv.date.to_formatted_s(:rfc822)}.  Did you go?" %></p>
+  <%= link_to 'Check event', event_show_path(inv)  %>
+  </div>
+</div>
+<% end %>
+<% end %>
+</div>
