+++ app/views/users/_users.html.erb
@@ -1,0 +1,12 @@
+<div class="p-4">
+<% @users.each do |user| %>
+<div class="card mb-2 " style="width: 18rem;">
+  <div class="card-body">
+  <h5 class="card-title"><%= "#{user.name} " %> </h5>
+  <p class="card-text"> <%= "created at #{user.created_at}." %></p>
+  <%= link_to 'Check user', user_show_path(user)  %>
+  </div>
+</div>
+<% end %>
+
+</div>
